sync-admin/requirements.txt
fastapi==0.115.12
uvicorn==0.30.6
sqlalchemy==2.0.36
pydantic==2.9.2
pydantic-settings==2.6.1
jinja2==3.1.4
python-multipart==0.0.10
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
PyJWT==2.9.0
psycopg2-binary==2.9.10
itsdangerous==2.2.0
httpx==0.27.2
